rename: components 폴더에 main, postdetail 폴더 생성 후 기존 파일들 분류
