+++ screen/Post.jsx
@@ -21,6 +21,8 @@ const Post = ({ navigation: { goBack } }) => {
     createdAt: Date.now(),
     userName: auth.currentUser.displayName,
     userId: auth.currentUser?.uid,
+    likes: 0,
+    liked: false,
   };
 
   const { isLoading, mutate: add } = useMutation(
